MCP/mobile_agent_mcp/requirements.txt
mcp
fastmcp
Pillow
requests